Move generate_otu_fasta() to virtool.otus.db
