+++ virtool/downloads/db.py
@@ -3,7 +3,7 @@ Functions for working with the database to provide file downloads. All functions
 child models.
 
 """
-from typing import Tuple, Optional
+from typing import Tuple
 
 import virtool.downloads.utils
 import virtool.errors
@@ -62,38 +62,6 @@ async def generate_sequence_fasta(db, sequence_id: str) -> Tuple[str, str]:
     )
 
     return virtool.downloads.utils.format_fasta_filename(otu_name, isolate_name, sequence["_id"]), fasta
-
-
-async def generate_otu_fasta(db, otu_id: str) -> Tuple[Optional[str], Optional[str]]:
-    """
-    Generate a FASTA filename and body for the sequences associated with the otu identified by the passed
-    ``otu_id``.
-
-    :param db: the application database client
-    :param otu_id: the id of the otu whose sequences should be expressed in FASTA format
-    :return: as FASTA filename and body
-
-    """
-    otu = await db.otus.find_one(otu_id, ["name", "isolates"])
-
-    fasta = list()
-
-    for isolate in otu["isolates"]:
-        sequences = await db.sequences.find({"otu_id": otu_id, "isolate_id": isolate["id"]}, ["sequence"]).to_list(None)
-        if sequences:
-            for sequence in sequences:
-                fasta.append(virtool.downloads.utils.format_fasta_entry(
-                    otu["name"],
-                    virtool.otus.utils.format_isolate_name(isolate),
-                    sequence["_id"],
-                    sequence["sequence"]
-                ))
-        else:
-            return None, None
-
-    fasta = "\n".join(fasta)
-
-    return virtool.downloads.utils.format_fasta_filename(otu["name"]), fasta
 
 
 async def get_otu_and_isolate_names(db, otu_id: str, isolate_id: str) -> Tuple[str, str]:
